--- forward-message-mailgun/tests/blank.test.js
@@ -1,9 +1,0 @@
-const handler = require('../functions/blank').handler;
-
-describe('forward-message-mailgun function template', () => {
-  it('Calls callback with empty JSON', () => {
-    const callback = jest.fn();
-    handler({}, {}, callback);
-    expect(callback).toHaveBeenCalledWith(null, {});
-  });
-});
